+++ deepinv/models/__init__.py
@@ -26,6 +26,7 @@ from .gan import (
     DCGANGenerator,
     DCGANDiscriminator,
     SkipConvDiscriminator,
+    UNetDiscriminatorSN,
 )
 from .complex import to_complex_denoiser
 from .dynamic import TimeAgnosticNet, TimeAveragingNet