+++ deepinv/training/adversarial.py
@@ -5,14 +5,14 @@ from deepinv.training.trainer import Trainer
 class AdversarialOptimizer:
     def __init__(*args, **kwargs):
         raise DeprecationWarning(
-            "AdversarialOptimizer is deprecated. See https://deepinv.github.io/deepinv/auto_examples/adversarial-learning/demo_gan_imaging.html for how to train adversarial networks in DeepInverse."
+            "AdversarialOptimizer is deprecated. See https://deepinv.github.io/deepinv/auto_examples/models/demo_gan_imaging.html for how to train adversarial networks in DeepInverse."
         )
 
 
 class AdversarialScheduler(AdversarialOptimizer):
     def __init__(*args, **kwargs):
         raise DeprecationWarning(
-            "AdversarialScheduler is deprecated. See https://deepinv.github.io/deepinv/auto_examples/adversarial-learning/demo_gan_imaging.html for how to train adversarial networks in DeepInverse."
+            "AdversarialScheduler is deprecated. See https://deepinv.github.io/deepinv/auto_examples/models/demo_gan_imaging.html for how to train adversarial networks in DeepInverse."
         )
 
 
@@ -20,5 +20,5 @@ class AdversarialScheduler(AdversarialOptimizer):
 class AdversarialTrainer(Trainer):
     def __init__(*args, **kwargs):
         raise DeprecationWarning(
-            "AdversarialTrainer is deprecated. See https://deepinv.github.io/deepinv/auto_examples/adversarial-learning/demo_gan_imaging.html for how to train adversarial networks in DeepInverse."
+            "AdversarialTrainer is deprecated. See https://deepinv.github.io/deepinv/auto_examples/models/demo_gan_imaging.html for how to train adversarial networks in DeepInverse."
         )